mbedtls: use more size_t to sync up with crypto.h
Ref: 5a96f494ee0b00282afb2db2e091246fc5e1774a #846 #879

Fixes #1053
Closes #1054

+++ src/mbedtls.h
@@ -478,12 +478,12 @@ _libssh2_mbedtls_cipher_dtor(_libssh2_cipher_ctx *ctx);
 int
 _libssh2_mbedtls_hash_init(mbedtls_md_context_t *ctx,
                            mbedtls_md_type_t mdtype,
-                           const unsigned char *key, unsigned long keylen);
+                           const unsigned char *key, size_t keylen);
 
 int
 _libssh2_mbedtls_hash_final(mbedtls_md_context_t *ctx, unsigned char *hash);
 int
-_libssh2_mbedtls_hash(const unsigned char *data, unsigned long datalen,
+_libssh2_mbedtls_hash(const unsigned char *data, size_t datalen,
                       mbedtls_md_type_t mdtype, unsigned char *hash);
 
 _libssh2_bn *
@@ -526,9 +526,8 @@ _libssh2_mbedtls_rsa_new_private_frommemory(libssh2_rsa_ctx **rsa,
 int
 _libssh2_mbedtls_rsa_sha1_verify(libssh2_rsa_ctx *rsa,
                                  const unsigned char *sig,
-                                 unsigned long sig_len,
-                                 const unsigned char *m,
-                                 unsigned long m_len);
+                                 size_t sig_len,
+                                 const unsigned char *m, size_t m_len);
 int
 _libssh2_mbedtls_rsa_sha1_sign(LIBSSH2_SESSION *session,
                                libssh2_rsa_ctx *rsa,
@@ -540,8 +539,8 @@ int
 _libssh2_mbedtls_rsa_sha2_verify(libssh2_rsa_ctx * rsactx,
                                  size_t hash_len,
                                  const unsigned char *sig,
-                                 unsigned long sig_len,
-                                 const unsigned char *m, unsigned long m_len);
+                                 size_t sig_len,
+                                 const unsigned char *m, size_t m_len);
 int
 _libssh2_mbedtls_rsa_sha2_sign(LIBSSH2_SESSION *session,
                                libssh2_rsa_ctx *rsa,
